# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Langsungmakan API",
        "contact": {
            "name": "Our",
            "url": "https://www.langsungmakan.com"
        },
        "version": "1.0"
    },
    "host": "wshgay.app",
    "basePath": "",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/v1/itag": {
            "get": {
                "tags": [
                    "itag"
                ],
                "summary": "Retrieves the collection of Itag resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itag collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/itag"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/itag/{id}": {
            "put": {
                "tags": [
                    "itag"
                ],
                "summary": "Edit itag resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "itag collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/itag"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "itag"
                ],
                "summary": "Removes the Itag resource.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Itag resource deleted."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            }
        },
        "/api/v1/item": {
            "get": {
                "tags": [
                    "item"
                ],
                "summary": "Retrieves the collection of Item resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/item"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/item/{id}": {
            "put": {
                "tags": [
                    "item"
                ],
                "summary": "Edit item resources. name,price,stock",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "item collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/item"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            },
            "delete": {
                "tags": [
                    "item"
                ],
                "summary": "Removes the Item resource.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Item resource deleted."
                    },
                    "401": {
                        "description": "Unauthorized action."
                    },
                    "404": {
                        "description": "Resource not found."
                    }
                }
            }
        },
...
        "/api/v1/Shop": {
            "get": {
                "tags": [
                    "shop"
                ],
                "summary": "Retrieves the collection of Shop resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Shop collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/shop"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
        "/api/v1/Shop/{id}": {
            "put": {
                "tags": [
                    "shop"
                ],
                "summary": "Edit shops resources.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "shop collection.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/shop"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized action."
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "itag": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "item": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer"
                },
                "stock": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "shop": {
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32"
                },
                "name": {
                    "type": "string"
                },
                "is_open": {
                    "type": "integer"
                },
                "open_at": {
                    "type": "string"
                },
                "close_at": {
                    "type": "string"
                },
                "shop_location_id": {
                    "type": "integer"
                },
                "shop_owner_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    }
}